backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# AI & ML
openai==1.10.0
tiktoken==0.5.2
anthropic==0.8.1

# Vector Search & Embeddings
faiss-cpu==1.7.4
numpy==1.26.3
scikit-learn==1.4.0
hdbscan==0.8.33

# HTTP & API Clients
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
feedparser==6.0.10
beautifulsoup4==4.12.3
lxml==5.1.0
newspaper3k==0.2.8

# Logging & Monitoring
loguru==0.7.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
